refactor (move directory) : model -> domain 패키지로 이동
