+++ src/main/java/racingcar/domain/Winner.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 import java.util.Collections;
 import java.util.PriorityQueue;
