+++ src/test/java/racingcar/model/CarsTest.java
@@ -3,6 +3,8 @@ package racingcar.model;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.domain.Cars;
 import racingcar.policy.fake.SuccessMovingPolicy;
 
 import java.util.ArrayList;
