+++ src/test/java/racingcar/model/RacingGameTest.java
@@ -3,6 +3,9 @@ package racingcar.model;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.controller.RacingGame;
+import racingcar.domain.Car;
+import racingcar.domain.RacingGameSetting;
 import racingcar.policy.fake.SuccessMovingPolicy;
 
 import static org.assertj.core.api.Assertions.assertThat;
