+++ src/test/java/racingcar/model/RoundTest.java
@@ -3,6 +3,7 @@ package racingcar.model;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Round;
 
 import static org.assertj.core.api.Assertions.assertThat;
