+++ src/test/java/racingcar/model/WinnerTest.java
@@ -3,6 +3,9 @@ package racingcar.model;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.domain.Cars;
+import racingcar.domain.Winner;
 
 import java.util.ArrayList;
 import java.util.List;
